+++ frontend/src/services/api/RentApiService.ts
@@ -1,16 +1,12 @@
-import { RentList } from '../../types/rent';
+import { Rent } from '../../types/rent';
 import { HttpService } from './HttpService';
+import { ModelApiService } from './ModelApiService';
 
-export class RentApiService extends HttpService {
-    public static readonly markRentPath = (id: number, paid: boolean) => `/api/rents/${id}/mark-as-${paid ? 'paid' : 'unpaid'}`;
-    public static readonly getRentsByTenantIdPath = (tenantId: number) => `/api/tenants/${tenantId}/rents`;
+export class RentApiService extends ModelApiService<Rent, 'rent'> {
+    public apiRoute = '/api/rents';
 
-    async updateRent(path: string) {
-        return HttpService.post(path, {});
-    }
-
-    async getRents(path: string) {
-        return HttpService.get<RentList>(path);
+    async mark(id: number, paid: boolean) {
+        return HttpService.post(`${this.apiRoute}/${id}/mark-as-${paid ? 'paid' : 'unpaid'}`, {});
     }
 }